% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/division.R
\name{division}
\alias{division}
\title{division}
\usage{
division(x, y)
}
\arguments{
\item{x}{Dividendo}

\item{y}{Divisor}
}
\value{
El resultado de dividir x entre y
}
\description{
divide un número entre otro
}
