% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicacion.R
\name{multiplicacion}
\alias{multiplicacion}
\title{multiplicacion}
\usage{
multiplicacion(x, y)
}
\arguments{
\item{x}{Primer número}

\item{y}{Segundo número}
}
\value{
El resultado de multiplicar x por y
}
\description{
multiplica dos números
}
